+++ ui/src/app/components/apps-panel/apps-panel-button.tsx
@@ -1,0 +1,74 @@
+import * as React from 'react';
+import { IconProp } from '@fortawesome/fontawesome-svg-core';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+import { Clickable } from 'app/components/buttons';
+import { FlexColumn } from 'app/components/flex';
+import colors from 'app/styles/colors';
+import { reactStyles } from 'app/utils';
+
+export const buttonStyles = reactStyles({
+  button: {
+    opacity: 1,
+    boxSizing: 'border-box',
+    height: 69,
+    width: 108,
+    border: '1px solid #979797',
+    borderRadius: 9,
+    alignItems: 'center',
+  },
+  disabledButton: {
+    opacity: 0.46,
+    cursor: 'not-allowed',
+    boxSizing: 'border-box',
+    height: 69,
+    width: 108,
+    border: '1px solid #979797',
+    borderRadius: 9,
+    alignItems: 'center',
+  },
+  buttonText: {
+    color: colors.primary,
+    fontFamily: 'Montserrat',
+    fontWeight: 'bold',
+    letterSpacing: 0,
+    lineHeight: '15px',
+  },
+  disabledButtonText: {
+    color: colors.secondary,
+    fontFamily: 'Montserrat',
+    fontWeight: 'bold',
+    letterSpacing: 0,
+    lineHeight: '15px',
+  },
+  buttonIcon: {
+    height: 23,
+    padding: '0.7em',
+  },
+});
+
+interface Props {
+  onClick: Function;
+  icon: IconProp;
+  buttonText: string;
+  disabled?: boolean;
+}
+export const AppsPanelButton = (props: Props) => {
+  const { disabled, onClick, icon, buttonText } = props;
+  return (
+    <Clickable {...{ disabled, onClick }} style={{ padding: '0.5em' }}>
+      <FlexColumn
+        style={disabled ? buttonStyles.disabledButton : buttonStyles.button}
+      >
+        <FontAwesomeIcon {...{ icon }} style={buttonStyles.buttonIcon} />
+        <div
+          style={
+            disabled ? buttonStyles.disabledButtonText : buttonStyles.buttonText
+          }
+        >
+          {buttonText}
+        </div>
+      </FlexColumn>
+    </Clickable>
+  );
+};
