+++ ui/src/app/components/apps-panel/new-notebook-button.tsx
@@ -1,17 +1,14 @@
 import * as React from 'react';
 import { useEffect, useState } from 'react';
 import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import { Workspace } from 'generated/fetch';
 
-import { Clickable } from 'app/components/buttons';
-import { FlexColumn } from 'app/components/flex';
 import { NewNotebookModal } from 'app/pages/analysis/new-notebook-modal';
 import { dropNotebookFileSuffix } from 'app/pages/analysis/util';
 import { notebooksApi } from 'app/services/swagger-fetch-clients';
 
-import { buttonStyles } from './utils';
+import { AppsPanelButton } from './apps-panel-button';
 
 export const NewNotebookButton = (props: { workspace: Workspace }) => {
   const { workspace } = props;
@@ -38,18 +35,11 @@ export const NewNotebookButton = (props: { workspace: Workspace }) => {
           }}
         />
       )}
-      <Clickable
-        style={{ padding: '0.5em' }}
+      <AppsPanelButton
         onClick={() => setShowModal(true)}
-      >
-        <FlexColumn style={buttonStyles.button}>
-          <FontAwesomeIcon
-            icon={faPlusCircle}
-            style={buttonStyles.buttonIcon}
-          />
-          <div style={buttonStyles.buttonText}>Create New</div>
-        </FlexColumn>
-      </Clickable>
+        icon={faPlusCircle}
+        buttonText='Create New'
+      />
     </>
   );
 };
